+++ NEWS.md
@@ -12,6 +12,7 @@
 * Add `location` argument to `az_resource_group$create_resource` method, rather than hardcoding it to the resgroup location.
 * Add `wait` argument when creating a new resource, similar to deploying a template, since some resources will return before provisioning is complete. Defaults to `FALSE` for backward compatibility.
 * Initialise AzureToken objects with an empty string as password instead of `NULL` when using device code flow; required by httr 1.4.0's stricter input checking.
+* Export `is_azure_token`.
 
 # AzureRMR 1.0.0